@@ -11,10 +11,6 @@
   "dependencies": {
     "@ducanh2912/next-pwa": "^10.2.6",
     "@headlessui/react": "^1.7.19",
-    "@nextui-org/modal": "^2.0.33",
-    "@nextui-org/react": "^2.3.6",
-    "@nextui-org/system": "^2.1.2",
-    "@nextui-org/theme": "^2.2.3",
     "@nivo/core": "^0.86.0",
     "@nivo/line": "^0.86.0",
     "@radix-ui/react-icons": "^1.3.0",
@@ -33,8 +29,8 @@
     "lucide-react": "^0.376.0",
     "next": "^14.2.3",
     "next-pwa": "^5.6.0",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-responsive": "^10.0.0",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7"
